refactor: Rename 'SearchAndSortBar' component as "SearchBar'
