+++ src/sort/SortPopover.tsx
@@ -16,7 +16,7 @@ import {
   Checkbox,
 } from "@material-ui/core";
 import { BaseTextButton, theme, BasePopover, icons } from "@gliff-ai/style";
-import { getLabelsFromKeys, MetadataLabel } from "@/search/SearchAndSortBar";
+import { getLabelsFromKeys, MetadataLabel } from "@/search/SearchBar";
 import { tooltips } from "@/components/Tooltips";
 
 const useStyles = makeStyles({
